refactor: extract shared build logic into dedicated module
- Move build_solana_program_internal from lib.rs to build_internal.rs
- Add cargo clean step before building to ensure fresh artifacts
- Update function calls in anchor and pinocchio modules to use new path
- Gate build_internal module with appropriate feature flags
- Remove redundant documentation reference in anchor build function
- Improve file extension checking with is_some_and for clarity

This refactoring improves code organization by separating build utilities
from the main library interface while enhancing build reliability.
